+++ lib/mirrmaid/synchronizer.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 
 # SPDX-License-Identifier: GPL-3.0-or-later
-# Copyright 2009-2018 John Florian <jflorian@doubledog.org>
+# Copyright 2009-2020 John Florian <jflorian@doubledog.org>
 #
 # This file is part of mirrmaid.
 
@@ -12,8 +12,12 @@ a perfect target replica of a source directory structure.  To ensure that only
 one synchronizer is working on a target replica at a time, advisory locking is
 utilized.
 """
+import errno
 import logging
 import os
+from subprocess import Popen
+from threading import Thread, Timer
+from time import sleep
 
 from doubledog.asynchronous import AsynchronousStreamingSubprocess
 from doubledog.lock import LockException, LockFile
@@ -21,17 +25,19 @@ from doubledog.lock import LockException, LockFile
 from mirrmaid.constants import *
 
 __author__ = """John Florian <jflorian@doubledog.org>"""
-__copyright__ = """Copyright 2009-2018 John Florian"""
+__copyright__ = """Copyright 2009-2020 John Florian"""
+
+STOP_TIMEOUT = 30
 
 
-class Synchronizer(object):
+class Synchronizer(Thread):
     """
-    This is effectively a Python wrapper around the venerable rsync, but made
-    suitable for mirrmaid use.  The options and arguments for rsync are mostly
-    derived from the mirrmaid configuration file.  A Synchronizer object works
-    on a single mirror.  Multiple Synchronizers may run concurrently provided
-    they are operating on distinct mirrors.  This is enforced via lock-files
-    on a per-mirror basis.
+    A thread to wrap around the venerable rsync, but made suitable for
+    mirrmaid use.  The options and arguments for rsync are mostly derived from
+    the mirrmaid configuration file.  A Synchronizer object works on a single
+    mirror.  Multiple Synchronizers may run concurrently provided they are
+    operating on distinct mirrors.  This is enforced via lock-files on
+    a per-mirror basis.
     """
 
     def __init__(self, default_conf, mirror_conf):
@@ -41,12 +47,13 @@ class Synchronizer(object):
         This will happen according to the default and mirror-specific
         sections of the configuration file..
         """
+        super().__init__()
         self.default_conf = default_conf
         self.mirror_conf = mirror_conf
-        self.log = logging.getLogger(
-            'mirrmaid.{0}'.format(self.mirror_conf.mirror_name)
-        )
+        self.log = logging.getLogger(f'mirrmaid.{self.mirror_conf.mirror_name}')
         self.lock_file = LockFile(self._lock_name, pid=os.getpid())
+        self.name = self.mirror_conf.mirror_name
+        self._subprocess = None
 
     @property
     def _lock_name(self) -> str:
@@ -89,7 +96,7 @@ class Synchronizer(object):
         return self.default_conf.get_list('rsync_options')
 
     @property
-    def _source(self) -> str:
+    def _source_uri(self) -> str:
         """
         :return:
             The fully-qualified rsync URI for the source of the directory
@@ -101,7 +108,7 @@ class Synchronizer(object):
         return source
 
     @property
-    def _target(self) -> str:
+    def _target_uri(self) -> str:
         """
         :return:
             The fully-qualified rsync URI for the target target of the mirroring
@@ -125,34 +132,21 @@ class Synchronizer(object):
         try:
             self.lock_file.exclusive_lock()
         except LockException:
-            self.log.info(
-                '{!r} already locked by another process'.format(
-                    self.lock_file.name,
-                )
-            )
+            self.log.info('%r already locked by another process',
+                          self.lock_file.name)
             return False
         else:
-            self.log.info(
-                'gained exclusive-lock on {!r}'.format(
-                    self.lock_file.name,
-                )
-            )
+            self.log.info('gained exclusive-lock on %r', self.lock_file.name)
             return True
 
     def _unlock_replica(self):
         """Release the lock on the target replica."""
         try:
             self.lock_file.unlock(delete_file=True)
-            self.log.info(
-                'released exclusive-lock on {!r}'.format(self.lock_file.name)
-            )
+            self.log.info('released exclusive-lock on %r', self.lock_file.name)
         except OSError as e:
-            self.log.error(
-                'failed to remove lock-file: {!r} because:\n{}'.format(
-                    self.lock_file.name,
-                    e,
-                )
-            )
+            self.log.error('failed to remove lock-file: %r because:\n%s',
+                           self.lock_file.name, e)
 
     def _update_replica(self) -> int:
         """
@@ -168,32 +162,71 @@ class Synchronizer(object):
         """
         self.log.info('mirror synchronization started')
         cmd = (
-            [RSYNC]
-            + self._rsync_options
-            + self._rsync_includes
-            + self._rsync_excludes
+                [RSYNC]
+                + self._rsync_options
+                + self._rsync_includes
+                + self._rsync_excludes
         )
-        cmd.append(self._source)
-        cmd.append(self._target)
-        self.log.debug('spawning {!r}'.format(cmd))
-        self.log.debug('AKA      {0}'.format(' '.join(cmd)))
-        process = AsynchronousStreamingSubprocess(cmd)
-        self.log.info('rsync pid={!r}'.format(process.pid))
-        exit_code = process.collect(self.log.info, self.log.error)
+        cmd.append(self._source_uri)
+        cmd.append(self._target_uri)
+        self.log.debug('spawning %r', cmd)
+        self.log.debug('AKA      %s', ' '.join(cmd))
+        self._subprocess = AsynchronousStreamingSubprocess(cmd)
+        self.log.info('rsync pid=%r', self._subprocess.pid)
+        exit_code = self._subprocess.collect(self.log.info, self.log.error)
         if exit_code < 0:
-            self.log.warn(
-                'rsync terminated; caught signal {!r}'.format(-exit_code)
-            )
+            self.log.warning('rsync terminated; caught signal %r', -exit_code)
         else:
             level = [logging.INFO, logging.DEBUG][exit_code == os.EX_OK]
-            self.log.log(level, 'rsync exit code={!r}'.format(exit_code))
+            self.log.log(level, 'rsync exit code=%r', exit_code)
         self.log.info('mirror synchronization finished')
         return exit_code
 
+    @property
+    def is_running(self) -> bool:
+        """
+        :returns:
+            ``True`` iff the rsync subprocess is currently running.
+        """
+        if self._subprocess and self._subprocess.pid:
+            try:
+                os.kill(self._subprocess.pid, 0)
+            except ProcessLookupError:
+                pass
+            else:
+                return True
+        return False
+
     def run(self):
         """Acquire a lock and if successful, update the target replica."""
+        self.log.info('starting thread')
         if self._lock_replica():
             try:
                 self._update_replica()
             finally:
                 self._unlock_replica()
+
+    def stop(self):
+        """Force termination of the rsync subprocess."""
+
+        def halt(msg, method):
+            self.log.info('%s %s', msg, self)
+            try:
+                method()
+            except OSError as e:
+                if e.errno != errno.ESRCH:  # no such process
+                    raise
+
+        if self._subprocess:
+            p: Popen = self._subprocess.process
+            halt('stopping', p.terminate)
+            t = Timer(STOP_TIMEOUT, halt, ('killing', p.kill))
+            t.start()
+            # exit quickly when possible
+            while t.is_alive():
+                if p.poll() is None:
+                    sleep(1)
+                else:
+                    t.cancel()
+                    self.log.info('%s stopped gracefully', self)
+                    break